feat: renamed OneUI.startCssVarsPonyfill to OneUI.applyCssVarsPonyfill()
Initial idea was to call OneUI.init() once on page load (that used to call
OneUI.startCssVarsPonyfill underthegood). But with code splitting
`cssVarsPonyfill` should be appled after any route navigation.
So it should be exposed with a more meaningfull name.

BREAKING CHANGE: if you used `OneUI.startCssVarsPonyfill()` directly in your app
rename it to `OneUI.applyCssVarsPonyfill()`.