feat(SelectButtonGroup): add support for controlled behavior and improve API (#624)
This adds support for using this component with controlled data while preserving the possibility to use it in an uncontrolled way. Following the API changes, (default) values are meant to be defined on `SelectButtonGroup`. Use `value` in a controlled implementation, or `defaultValue` when uncontrolled. Context (prop is renamed to `context`) can be defined both on `SelectButtonGroup` level (applies to all children) or on each `SelectButton` individually.

BREAKING CHANGE: The API for these components is changed because of the improvements and changes in behavior that are implemented.